chore(types) Move more types out of types/index
Move project, release and group types out. There is one circular import
between Commit and Release. A potential solution for that cycle is to
have Commit in release module but it feels odd for it to be there when
repository and other source code related types are in integrations.
