+++ static/app/components/events/interfaces/threads/threadSelector/findBestThread.tsx
@@ -1,4 +1,4 @@
-import {Thread} from 'sentry/types/events';
+import {Thread} from 'sentry/types';
 
 function findBestThread(threads: Array<Thread>) {
   // search the entire threads list for a crashed thread with stack trace
