+++ static/app/components/events/interfaces/threadsV2.tsx
@@ -5,10 +5,15 @@ import isNil from 'lodash/isNil';
 import Pill from 'sentry/components/pill';
 import Pills from 'sentry/components/pills';
 import {t} from 'sentry/locale';
-import {Frame, PlatformType, Project} from 'sentry/types';
-import {Event} from 'sentry/types/event';
-import {Thread} from 'sentry/types/events';
-import {STACK_TYPE, STACK_VIEW} from 'sentry/types/stacktrace';
+import {
+  Event,
+  Frame,
+  PlatformType,
+  Project,
+  STACK_TYPE,
+  STACK_VIEW,
+  Thread,
+} from 'sentry/types';
 
 import TraceEventDataSection from '../traceEventDataSection';
 import {DisplayOption} from '../traceEventDataSection/displayOptions';
