--- static/app/types/events.tsx
@@ -1,10 +1,0 @@
-import {RawStacktrace, StacktraceType} from './stacktrace';
-
-export interface Thread {
-  id: number;
-  crashed: boolean;
-  current: boolean;
-  rawStacktrace: RawStacktrace;
-  stacktrace: StacktraceType | null;
-  name?: string | null;
-}
