+++ static/app/types/index.tsx
@@ -6,24 +6,23 @@ import {FocusTrap} from 'focus-trap';
 
 import exportGlobals from 'sentry/bootstrap/exportGlobals';
 import {getInterval} from 'sentry/components/charts/utils';
-import {SymbolicatorStatus} from 'sentry/components/events/interfaces/types';
 import {DEFAULT_RELATIVE_PERIODS} from 'sentry/constants';
-import {PlatformKey} from 'sentry/data/platformCategories';
 
-import {Actor, DateString, Scope, TimeseriesValue} from './core';
-import {DynamicSamplingRules} from './dynamicSampling';
-import {Event} from './event';
-import {Plugin, PullRequest, Repository} from './integrations';
+import {DateString, Scope} from './core';
 import {Member, Organization, Team} from './organization';
-import {RawStacktrace, StackTraceMechanism, StacktraceType} from './stacktrace';
-import {AvatarUser, User} from './user';
+import {User} from './user';
 
 export * from './auth';
 export * from './core';
+export * from './event';
+export * from './group';
 export * from './integrations';
 export * from './onboarding';
 export * from './organization';
+export * from './project';
 export * from './relay';
+export * from './release';
+export * from './stacktrace';
 export * from './user';
 
 export enum SentryInitRenderReactComponent {
@@ -134,198 +133,6 @@ export type PipelineInitialData = {
   props: Record<string, any>;
 };
 
-// Minimal project representation for use with avatars.
-export type AvatarProject = {
-  slug: string;
-  platform?: PlatformKey;
-  id?: string | number;
-};
-
-export type Project = {
-  id: string;
-  dateCreated: string;
-  isMember: boolean;
-  teams: Team[];
-  features: string[];
-  organization: Organization;
-
-  isBookmarked: boolean;
-  isInternal: boolean;
-  hasUserReports?: boolean;
-  hasAccess: boolean;
-  hasSessions: boolean;
-  firstEvent: 'string' | null;
-  firstTransactionEvent: boolean;
-  subjectTemplate: string;
-  digestsMaxDelay: number;
-  digestsMinDelay: number;
-  environments: string[];
-  eventProcessing: {
-    symbolicationDegraded: boolean;
-  };
-
-  // XXX: These are part of the DetailedProject serializer
-  dynamicSampling: {
-    next_id: number;
-    rules: DynamicSamplingRules;
-  } | null;
-  plugins: Plugin[];
-  processingIssues: number;
-  relayPiiConfig: string;
-  groupingConfig: string;
-  latestDeploys?: Record<string, Pick<Deploy, 'dateFinished' | 'version'>> | null;
-  builtinSymbolSources?: string[];
-  symbolSources?: string;
-  stats?: TimeseriesValue[];
-  transactionStats?: TimeseriesValue[];
-  latestRelease?: Release;
-  options?: Record<string, boolean | string>;
-  sessionStats?: {
-    currentCrashFreeRate: number | null;
-    previousCrashFreeRate: number | null;
-    hasHealthData: boolean;
-  };
-} & AvatarProject;
-
-export type MinimalProject = Pick<Project, 'id' | 'slug' | 'platform'>;
-
-// Response from project_keys endpoints.
-export type ProjectKey = {
-  id: string;
-  name: string;
-  label: string;
-  public: string;
-  secret: string;
-  projectId: string;
-  isActive: boolean;
-  rateLimit: {
-    window: string;
-    count: number;
-  } | null;
-  dsn: {
-    secret: string;
-    public: string;
-    csp: string;
-    security: string;
-    minidump: string;
-    unreal: string;
-    cdn: string;
-  };
-  browserSdkVersion: string;
-  browserSdk: {
-    choices: [key: string, value: string][];
-  };
-  dateCreated: string;
-};
-
-export type Health = {
-  totalUsers: number;
-  totalUsers24h: number | null;
-  totalProjectUsers24h: number | null;
-  totalSessions: number;
-  totalSessions24h: number | null;
-  totalProjectSessions24h: number | null;
-  crashFreeUsers: number | null;
-  crashFreeSessions: number | null;
-  stats: HealthGraphData;
-  sessionsCrashed: number;
-  sessionsErrored: number;
-  adoption: number | null;
-  sessionsAdoption: number | null;
-  hasHealthData: boolean;
-  durationP50: number | null;
-  durationP90: number | null;
-};
-
-export type HealthGraphData = Record<string, TimeseriesValue[]>;
-
-export type TeamWithProjects = Team & {projects: Project[]};
-
-export type TreeLabelPart =
-  | string
-  | {
-      function?: string;
-      package?: string;
-      type?: string;
-      classbase?: string;
-      filebase?: string;
-      datapath?: (string | number)[];
-      // is_sentinel is no longer being used,
-      // but we will still assess whether we will use this property in the near future.
-      is_sentinel?: boolean;
-      is_prefix?: boolean;
-    };
-
-// This type is incomplete
-export type EventMetadata = {
-  value?: string;
-  message?: string;
-  directive?: string;
-  type?: string;
-  title?: string;
-  uri?: string;
-  filename?: string;
-  origin?: string;
-  function?: string;
-  stripped_crash?: boolean;
-  current_tree_label?: TreeLabelPart[];
-  finest_tree_label?: TreeLabelPart[];
-  current_level?: number;
-  display_title_with_tree_label?: boolean;
-};
-
-// endpoint: /api/0/issues/:issueId/attachments/?limit=50
-export type IssueAttachment = {
-  id: string;
-  dateCreated: string;
-  headers: Object;
-  mimetype: string;
-  name: string;
-  sha1: string;
-  size: number;
-  type: string;
-  event_id: string;
-};
-
-// endpoint: /api/0/projects/:orgSlug/:projSlug/events/:eventId/attachments/
-export type EventAttachment = Omit<IssueAttachment, 'event_id'>;
-
-export type EntryData = Record<string, any | Array<any>>;
-
-type EnableIntegrationSuggestion = {
-  type: 'enableIntegration';
-  integrationName: string;
-  enables: Array<SDKUpdatesSuggestion>;
-  integrationUrl?: string | null;
-};
-
-export type UpdateSdkSuggestion = {
-  type: 'updateSdk';
-  sdkName: string;
-  newSdkVersion: string;
-  enables: Array<SDKUpdatesSuggestion>;
-  sdkUrl?: string | null;
-};
-
-type ChangeSdkSuggestion = {
-  type: 'changeSdk';
-  newSdkName: string;
-  enables: Array<SDKUpdatesSuggestion>;
-  sdkUrl?: string | null;
-};
-
-export type SDKUpdatesSuggestion =
-  | EnableIntegrationSuggestion
-  | UpdateSdkSuggestion
-  | ChangeSdkSuggestion;
-
-export type ProjectSdkUpdates = {
-  projectId: string;
-  sdkName: string;
-  sdkVersion: string;
-  suggestions: SDKUpdatesSuggestion[];
-};
-
 export type EventsStatsData = [number, {count: number; comparisonCount?: number}[]][];
 export type EventsGeoData = {'geo.country_code': string; count: number}[];
 
@@ -342,42 +149,6 @@ export type EventsStats = {
 export type MultiSeriesEventsStats = {
   [seriesName: string]: EventsStats;
 };
-
-export type Environment = {
-  id: string;
-  displayName: string;
-  name: string;
-
-  // XXX: Provided by the backend but unused due to `getUrlRoutingName()`
-  // urlRoutingName: string;
-};
-
-export type RecentSearch = {
-  id: string;
-  organizationId: string;
-  type: SavedSearchType;
-  query: string;
-  lastSeen: string;
-  dateCreated: string;
-};
-
-// XXX: Deprecated Sentry 9 attributes are not included here.
-export type SavedSearch = {
-  id: string;
-  type: SavedSearchType;
-  name: string;
-  query: string;
-  sort: string;
-  isGlobal: boolean;
-  isPinned: boolean;
-  isOrgCustom: boolean;
-  dateCreated: string;
-};
-
-export enum SavedSearchType {
-  ISSUE = 0,
-  EVENT = 1,
-}
 
 export type RelativePeriod = keyof typeof DEFAULT_RELATIVE_PERIODS;
 export type IntervalPeriod = ReturnType<typeof getInterval>;
@@ -445,375 +216,6 @@ export interface Config {
   };
 }
 
-// https://github.com/getsentry/relay/blob/master/relay-common/src/constants.rs
-// Note: the value of the enum on the frontend is plural,
-// but the value of the enum on the backend is singular
-export enum DataCategory {
-  DEFAULT = 'default',
-  ERRORS = 'errors',
-  TRANSACTIONS = 'transactions',
-  ATTACHMENTS = 'attachments',
-}
-
-export type EventType = 'error' | 'transaction' | 'attachment';
-
-export const DataCategoryName = {
-  [DataCategory.ERRORS]: 'Errors',
-  [DataCategory.TRANSACTIONS]: 'Transactions',
-  [DataCategory.ATTACHMENTS]: 'Attachments',
-};
-
-export enum EventOrGroupType {
-  ERROR = 'error',
-  CSP = 'csp',
-  HPKP = 'hpkp',
-  EXPECTCT = 'expectct',
-  EXPECTSTAPLE = 'expectstaple',
-  DEFAULT = 'default',
-  TRANSACTION = 'transaction',
-}
-
-export type InboxReasonDetails = {
-  until?: string | null;
-  count?: number | null;
-  window?: number | null;
-  user_count?: number | null;
-  user_window?: number | null;
-};
-
-export type InboxDetails = {
-  reason_details: InboxReasonDetails;
-  date_added?: string;
-  reason?: number;
-};
-
-export type SuggestedOwnerReason = 'suspectCommit' | 'ownershipRule';
-
-// Received from the backend to denote suggested owners of an issue
-export type SuggestedOwner = {
-  type: SuggestedOwnerReason;
-  owner: string;
-  date_added: string;
-};
-
-export enum GroupActivityType {
-  NOTE = 'note',
-  SET_RESOLVED = 'set_resolved',
-  SET_RESOLVED_BY_AGE = 'set_resolved_by_age',
-  SET_RESOLVED_IN_RELEASE = 'set_resolved_in_release',
-  SET_RESOLVED_IN_COMMIT = 'set_resolved_in_commit',
-  SET_RESOLVED_IN_PULL_REQUEST = 'set_resolved_in_pull_request',
-  SET_UNRESOLVED = 'set_unresolved',
-  SET_IGNORED = 'set_ignored',
-  SET_PUBLIC = 'set_public',
-  SET_PRIVATE = 'set_private',
-  SET_REGRESSION = 'set_regression',
-  CREATE_ISSUE = 'create_issue',
-  UNMERGE_SOURCE = 'unmerge_source',
-  UNMERGE_DESTINATION = 'unmerge_destination',
-  FIRST_SEEN = 'first_seen',
-  ASSIGNED = 'assigned',
-  UNASSIGNED = 'unassigned',
-  MERGE = 'merge',
-  REPROCESS = 'reprocess',
-  MARK_REVIEWED = 'mark_reviewed',
-}
-
-type GroupActivityBase = {
-  dateCreated: string;
-  id: string;
-  project: Project;
-  user?: null | User;
-  assignee?: string;
-  issue?: Group;
-};
-
-type GroupActivityNote = GroupActivityBase & {
-  type: GroupActivityType.NOTE;
-  data: {
-    text: string;
-  };
-};
-
-type GroupActivitySetResolved = GroupActivityBase & {
-  type: GroupActivityType.SET_RESOLVED;
-  data: Record<string, any>;
-};
-
-type GroupActivitySetUnresolved = GroupActivityBase & {
-  type: GroupActivityType.SET_UNRESOLVED;
-  data: Record<string, any>;
-};
-
-type GroupActivitySetPublic = GroupActivityBase & {
-  type: GroupActivityType.SET_PUBLIC;
-  data: Record<string, any>;
-};
-
-type GroupActivitySetPrivate = GroupActivityBase & {
-  type: GroupActivityType.SET_PRIVATE;
-  data: Record<string, any>;
-};
-
-type GroupActivitySetByAge = GroupActivityBase & {
-  type: GroupActivityType.SET_RESOLVED_BY_AGE;
-  data: Record<string, any>;
-};
-
-type GroupActivityUnassigned = GroupActivityBase & {
-  type: GroupActivityType.UNASSIGNED;
-  data: Record<string, any>;
-};
-
-type GroupActivityFirstSeen = GroupActivityBase & {
-  type: GroupActivityType.FIRST_SEEN;
-  data: Record<string, any>;
-};
-
-type GroupActivityMarkReviewed = GroupActivityBase & {
-  type: GroupActivityType.MARK_REVIEWED;
-  data: Record<string, any>;
-};
-
-type GroupActivityRegression = GroupActivityBase & {
-  type: GroupActivityType.SET_REGRESSION;
-  data: {
-    version?: string;
-  };
-};
-
-export type GroupActivitySetByResolvedInRelease = GroupActivityBase & {
-  type: GroupActivityType.SET_RESOLVED_IN_RELEASE;
-  data: {
-    version?: string;
-    current_release_version?: string;
-  };
-};
-
-type GroupActivitySetByResolvedInCommit = GroupActivityBase & {
-  type: GroupActivityType.SET_RESOLVED_IN_COMMIT;
-  data: {
-    commit: Commit;
-  };
-};
-
-type GroupActivitySetByResolvedInPullRequest = GroupActivityBase & {
-  type: GroupActivityType.SET_RESOLVED_IN_PULL_REQUEST;
-  data: {
-    pullRequest: PullRequest;
-  };
-};
-
-export type GroupActivitySetIgnored = GroupActivityBase & {
-  type: GroupActivityType.SET_IGNORED;
-  data: {
-    ignoreDuration?: number;
-    ignoreUntil?: string;
-    ignoreUserCount?: number;
-    ignoreUserWindow?: number;
-    ignoreWindow?: number;
-    ignoreCount?: number;
-  };
-};
-
-export type GroupActivityReprocess = GroupActivityBase & {
-  type: GroupActivityType.REPROCESS;
-  data: {
-    eventCount: number;
-    newGroupId: number;
-    oldGroupId: number;
-  };
-};
-
-type GroupActivityUnmergeDestination = GroupActivityBase & {
-  type: GroupActivityType.UNMERGE_DESTINATION;
-  data: {
-    fingerprints: Array<string>;
-    source?: {
-      id: string;
-      shortId: string;
-    };
-  };
-};
-
-type GroupActivityUnmergeSource = GroupActivityBase & {
-  type: GroupActivityType.UNMERGE_SOURCE;
-  data: {
-    fingerprints: Array<string>;
-    destination?: {
-      id: string;
-      shortId: string;
-    };
-  };
-};
-
-type GroupActivityMerge = GroupActivityBase & {
-  type: GroupActivityType.MERGE;
-  data: {
-    issues: Array<any>;
-  };
-};
-
-export type GroupActivityAssigned = GroupActivityBase & {
-  type: GroupActivityType.ASSIGNED;
-  data: {
-    assignee: string;
-    assigneeType: string;
-    user: Team | User;
-  };
-};
-
-export type GroupActivityCreateIssue = GroupActivityBase & {
-  type: GroupActivityType.CREATE_ISSUE;
-  data: {
-    provider: string;
-    location: string;
-    title: string;
-  };
-};
-
-export type GroupActivity =
-  | GroupActivityNote
-  | GroupActivitySetResolved
-  | GroupActivitySetUnresolved
-  | GroupActivitySetIgnored
-  | GroupActivitySetByAge
-  | GroupActivitySetByResolvedInRelease
-  | GroupActivitySetByResolvedInCommit
-  | GroupActivitySetByResolvedInPullRequest
-  | GroupActivityFirstSeen
-  | GroupActivityMerge
-  | GroupActivityReprocess
-  | GroupActivityUnassigned
-  | GroupActivityMarkReviewed
-  | GroupActivityUnmergeDestination
-  | GroupActivitySetPublic
-  | GroupActivitySetPrivate
-  | GroupActivityRegression
-  | GroupActivityUnmergeSource
-  | GroupActivityAssigned
-  | GroupActivityCreateIssue;
-
-export type Activity = GroupActivity;
-
-type GroupFiltered = {
-  count: string;
-  stats: Record<string, TimeseriesValue[]>;
-  lastSeen: string;
-  firstSeen: string;
-  userCount: number;
-};
-
-export type GroupStats = GroupFiltered & {
-  lifetime?: GroupFiltered;
-  filtered: GroupFiltered | null;
-  sessionCount?: string | null;
-  id: string;
-};
-
-export type BaseGroupStatusReprocessing = {
-  status: 'reprocessing';
-  statusDetails: {
-    pendingEvents: number;
-    info: {
-      dateCreated: string;
-      totalEvents: number;
-    } | null;
-  };
-};
-
-type BaseGroupStatusResolution = {
-  status: ResolutionStatus;
-  statusDetails: ResolutionStatusDetails;
-};
-
-export type GroupRelease = {
-  firstRelease: Release;
-  lastRelease: Release;
-};
-
-// TODO(ts): incomplete
-export type BaseGroup = {
-  id: string;
-  latestEvent: Event;
-  activity: GroupActivity[];
-  annotations: string[];
-  assignedTo: Actor;
-  culprit: string;
-  firstSeen: string;
-  hasSeen: boolean;
-  isBookmarked: boolean;
-  isUnhandled: boolean;
-  isPublic: boolean;
-  isSubscribed: boolean;
-  lastSeen: string;
-  level: Level;
-  logger: string;
-  metadata: EventMetadata;
-  numComments: number;
-  participants: User[];
-  permalink: string;
-  platform: PlatformKey;
-  pluginActions: any[]; // TODO(ts)
-  pluginContexts: any[]; // TODO(ts)
-  pluginIssues: any[]; // TODO(ts)
-  project: Project;
-  seenBy: User[];
-  shareId: string;
-  shortId: string;
-  tags: Pick<Tag, 'key' | 'name' | 'totalValues'>[];
-  title: string;
-  type: EventOrGroupType;
-  userReportCount: number;
-  subscriptionDetails: {disabled?: boolean; reason?: string} | null;
-  status: string;
-  inbox?: InboxDetails | null | false;
-  owners?: SuggestedOwner[] | null;
-} & GroupRelease;
-
-export type GroupReprocessing = BaseGroup & GroupStats & BaseGroupStatusReprocessing;
-export type GroupResolution = BaseGroup & GroupStats & BaseGroupStatusResolution;
-export type Group = GroupResolution | GroupReprocessing;
-export type GroupCollapseRelease = Omit<Group, keyof GroupRelease> &
-  Partial<GroupRelease>;
-
-export type GroupTombstone = {
-  id: string;
-  title: string;
-  culprit: string;
-  level: Level;
-  actor: AvatarUser;
-  metadata: EventMetadata;
-};
-
-export type ProcessingIssueItem = {
-  id: string;
-  type: string;
-  checksum: string;
-  numEvents: number;
-  data: {
-    // TODO(ts) This type is likely incomplete, but this is what
-    // project processing issues settings uses.
-    _scope: string;
-    image_arch: string;
-    image_uuid: string;
-    image_path: string;
-  };
-  lastSeen: string;
-};
-
-export type ProcessingIssue = {
-  project: string;
-  numIssues: number;
-  signedLink: string;
-  lastSeen: string;
-  hasMoreResolveableIssues: boolean;
-  hasIssues: boolean;
-  issuesProcessing: number;
-  resolveableIssues: number;
-  issues?: ProcessingIssueItem[];
-};
-
 export type AccessRequest = {
   id: string;
   team: Team;
@@ -851,140 +253,6 @@ export type ApiApplication = {
   privacyUrl: string | null;
   redirectUris: string[];
   termsUrl: string | null;
-};
-
-export type UserReport = {
-  id: string;
-  eventID: string;
-  issue: Group;
-  name: string;
-  event: {eventID: string; id: string};
-  user: User;
-  dateCreated: string;
-  comments: string;
-  email: string;
-};
-
-export type Release = BaseRelease &
-  ReleaseData & {
-    projects: ReleaseProject[];
-  };
-
-export type ReleaseWithHealth = BaseRelease &
-  ReleaseData & {
-    projects: Required<ReleaseProject>[];
-  };
-
-type ReleaseData = {
-  commitCount: number;
-  data: {};
-  lastDeploy?: Deploy;
-  deployCount: number;
-  lastEvent: string;
-  firstEvent: string;
-  lastCommit?: Commit;
-  authors: User[];
-  owner?: any; // TODO(ts)
-  newGroups: number;
-  versionInfo: VersionInfo;
-  fileCount: number | null;
-  currentProjectMeta: {
-    nextReleaseVersion: string | null;
-    prevReleaseVersion: string | null;
-    sessionsLowerBound: string | null;
-    sessionsUpperBound: string | null;
-    firstReleaseVersion: string | null;
-    lastReleaseVersion: string | null;
-  };
-  adoptionStages?: Record<
-    'string',
-    {
-      stage: string | null;
-      adopted: string | null;
-      unadopted: string | null;
-    }
-  >;
-};
-
-type BaseRelease = {
-  dateReleased: string;
-  url: string;
-  dateCreated: string;
-  version: string;
-  shortVersion: string;
-  ref: string;
-  status: ReleaseStatus;
-};
-
-export type CurrentRelease = {
-  environment: string;
-  firstSeen: string;
-  lastSeen: string;
-  release: Release;
-  stats: {
-    // 24h/30d is hardcoded in GroupReleaseWithStatsSerializer
-    '24h': TimeseriesValue[];
-    '30d': TimeseriesValue[];
-  };
-};
-export enum ReleaseStatus {
-  Active = 'open',
-  Archived = 'archived',
-}
-
-// XXX This should go to types/integrations but release needs to move first.
-export type Commit = {
-  id: string;
-  message: string | null;
-  dateCreated: string;
-  releases: BaseRelease[];
-  repository?: Repository;
-  author?: User;
-};
-
-// XXX This should go to types/integrations but Commit needs to move first.
-export type Committer = {
-  author: User;
-  commits: Commit[];
-};
-
-export type ReleaseProject = {
-  slug: string;
-  name: string;
-  id: number;
-  platform: PlatformKey;
-  platforms: PlatformKey[];
-  newGroups: number;
-  hasHealthData: boolean;
-  healthData?: Health;
-};
-
-export type ReleaseMeta = {
-  commitCount: number;
-  commitFilesChanged: number;
-  deployCount: number;
-  releaseFileCount: number;
-  version: string;
-  projects: ReleaseProject[];
-  versionInfo: VersionInfo;
-  released: string;
-};
-
-export type VersionInfo = {
-  buildHash: string | null;
-  description: string;
-  package: string | null;
-  version: {raw: string};
-};
-
-export type Deploy = {
-  id: string;
-  name: string;
-  url: string;
-  environment: string;
-  dateStarted: string;
-  dateFinished: string;
-  version: string;
 };
 
 export type SavedQueryVersions = 1 | 2;
@@ -1029,100 +297,6 @@ export type SavedQueryState = {
   isLoading: boolean;
 };
 
-export type Tag = {
-  name: string;
-  key: string;
-  values?: string[];
-  totalValues?: number;
-  predefined?: boolean;
-  isInput?: boolean;
-  /**
-   * How many values should be suggested in autocomplete.
-   * Overrides SmartSearchBar's `maxSearchItems` prop.
-   */
-  maxSuggestedValues?: number;
-};
-
-export type TagCollection = Record<string, Tag>;
-
-export type TagValue = {
-  count: number;
-  name: string;
-  value: string;
-  lastSeen: string;
-  key: string;
-  firstSeen: string;
-  query?: string;
-  email?: string;
-  username?: string;
-  identifier?: string;
-  ipAddress?: string;
-} & AvatarUser;
-
-type Topvalue = {
-  count: number;
-  firstSeen: string;
-  key: string;
-  lastSeen: string;
-  name: string;
-  value: string;
-  // Might not actually exist.
-  query?: string;
-};
-
-export type TagWithTopValues = {
-  topValues: Array<Topvalue>;
-  key: string;
-  name: string;
-  totalValues: number;
-  uniqueValues: number;
-  canDelete?: boolean;
-};
-
-export type Level = 'error' | 'fatal' | 'info' | 'warning' | 'sample';
-
-export type Meta = {
-  chunks: Array<ChunkType>;
-  len: number;
-  rem: Array<MetaRemark>;
-  err: Array<MetaError>;
-};
-
-export type MetaError = string | [string, any];
-export type MetaRemark = Array<string | number>;
-
-export type ChunkType = {
-  text: string;
-  type: string;
-  rule_id: string | number;
-  remark?: string | number;
-};
-
-export enum ResolutionStatus {
-  RESOLVED = 'resolved',
-  UNRESOLVED = 'unresolved',
-  IGNORED = 'ignored',
-}
-export type ResolutionStatusDetails = {
-  actor?: AvatarUser;
-  autoResolved?: boolean;
-  ignoreCount?: number;
-  // Sent in requests. ignoreUntil is used in responses.
-  ignoreDuration?: number;
-  ignoreUntil?: string;
-  ignoreUserCount?: number;
-  ignoreUserWindow?: number;
-  ignoreWindow?: number;
-  inCommit?: Commit;
-  inRelease?: string;
-  inNextRelease?: boolean;
-};
-
-export type UpdateResolutionStatus = {
-  status: ResolutionStatus;
-  statusDetails?: ResolutionStatusDetails;
-};
-
 export type SubscriptionDetails = {disabled?: boolean; reason?: string};
 
 export type Broadcast = {
@@ -1165,134 +339,6 @@ export type PlatformIntegration = {
   language: string;
   link: string | null;
   name: string;
-};
-
-export type EventGroupComponent = {
-  contributes: boolean;
-  hint: string | null;
-  id: string;
-  name: string | null;
-  values: EventGroupComponent[] | string[];
-};
-export type EventGroupingConfig = {
-  base: string | null;
-  changelog: string;
-  delegates: string[];
-  hidden: boolean;
-  id: string;
-  latest: boolean;
-  risk: number;
-  strategies: string[];
-};
-
-type EventGroupVariantKey = 'custom-fingerprint' | 'app' | 'default' | 'system';
-
-export enum EventGroupVariantType {
-  CUSTOM_FINGERPRINT = 'custom-fingerprint',
-  COMPONENT = 'component',
-  SALTED_COMPONENT = 'salted-component',
-}
-
-export type EventGroupVariant = {
-  description: string | null;
-  hash: string | null;
-  hashMismatch: boolean;
-  key: EventGroupVariantKey;
-  type: EventGroupVariantType;
-  values?: Array<string>;
-  client_values?: Array<string>;
-  matched_rule?: string;
-  component?: EventGroupComponent;
-  config?: EventGroupingConfig;
-};
-
-export type SourceMapsArchive = {
-  id: number;
-  type: 'release';
-  name: string;
-  date: string;
-  fileCount: number;
-};
-
-export type Artifact = {
-  dateCreated: string;
-  dist: string | null;
-  id: string;
-  name: string;
-  sha1: string;
-  size: number;
-  headers: {'Content-Type': string};
-};
-
-export type EventGroupInfo = Record<EventGroupVariantKey, EventGroupVariant>;
-
-// TODO(epurkhiser): objc and cocoa should almost definitely be moved into PlatformKey
-export type PlatformType = PlatformKey | 'objc' | 'cocoa';
-
-export type Frame = {
-  absPath: string | null;
-  colNo: number | null;
-  context: Array<[number, string]>;
-  errors: Array<any> | null;
-  filename: string | null;
-  function: string | null;
-  inApp: boolean;
-  instructionAddr: string | null;
-  lineNo: number | null;
-  module: string | null;
-  package: string | null;
-  platform: PlatformType | null;
-  rawFunction: string | null;
-  symbol: string | null;
-  symbolAddr: string | null;
-  trust: any | null;
-  vars: Record<string, any> | null;
-  symbolicatorStatus?: SymbolicatorStatus;
-  addrMode?: string;
-  origAbsPath?: string | null;
-  mapUrl?: string | null;
-  map?: string | null;
-  isSentinel?: boolean;
-  isPrefix?: boolean;
-  minGroupingLevel?: number;
-};
-
-export enum FrameBadge {
-  SENTINEL = 'sentinel',
-  PREFIX = 'prefix',
-  GROUPING = 'grouping',
-}
-
-/**
- * Note used in Group Activity and Alerts for users to comment
- */
-export type Note = {
-  /**
-   * Note contents (markdown allowed)
-   */
-  text: string;
-
-  /**
-   * Array of [id, display string] tuples used for @-mentions
-   */
-  mentions: [string, string][];
-};
-
-export type ExceptionValue = {
-  type: string;
-  value: string;
-  threadId: number | null;
-  stacktrace: StacktraceType | null;
-  rawStacktrace: RawStacktrace;
-  mechanism: StackTraceMechanism | null;
-  module: string | null;
-  frames?: Frame[] | null;
-};
-
-export type ExceptionType = {
-  excOmitted: any | null;
-  hasSystemFrames: boolean;
-  values?: Array<ExceptionValue>;
 };
 
 export enum UserIdentityCategory {
@@ -1348,7 +394,7 @@ export type ServerlessFunction = {
 };
 
 /**
- * Base type for series   style API response
+ * Base type for series style API response
  */
 export type SeriesApi = {
   intervals: string[];
@@ -1377,65 +423,3 @@ export enum SessionStatus {
   ERRORED = 'errored',
   CRASHED = 'crashed',
 }
-
-export enum ReleaseComparisonChartType {
-  CRASH_FREE_USERS = 'crashFreeUsers',
-  HEALTHY_USERS = 'healthyUsers',
-  ABNORMAL_USERS = 'abnormalUsers',
-  ERRORED_USERS = 'erroredUsers',
-  CRASHED_USERS = 'crashedUsers',
-  CRASH_FREE_SESSIONS = 'crashFreeSessions',
-  HEALTHY_SESSIONS = 'healthySessions',
-  ABNORMAL_SESSIONS = 'abnormalSessions',
-  ERRORED_SESSIONS = 'erroredSessions',
-  CRASHED_SESSIONS = 'crashedSessions',
-  SESSION_COUNT = 'sessionCount',
-  USER_COUNT = 'userCount',
-  ERROR_COUNT = 'errorCount',
-  TRANSACTION_COUNT = 'transactionCount',
-  FAILURE_RATE = 'failureRate',
-  SESSION_DURATION = 'sessionDuration',
-}
-
-export enum HealthStatsPeriodOption {
-  AUTO = 'auto',
-  TWENTY_FOUR_HOURS = '24h',
-}
-
-export type IssueOwnership = {
-  raw: string;
-  fallthrough: boolean;
-  dateCreated: string;
-  lastUpdated: string;
-  isActive: boolean;
-  autoAssignment: boolean;
-};
-
-export type KeyValueListData = {
-  key: string;
-  subject: string;
-  value?: React.ReactNode;
-  meta?: Meta;
-  subjectDataTestId?: string;
-  subjectIcon?: React.ReactNode;
-}[];
-
-// Response from ShortIdLookupEndpoint
-// /organizations/${orgId}/shortids/${query}/
-export type ShortIdResponse = {
-  organizationSlug: string;
-  projectSlug: string;
-  groupId: string;
-  group: Group;
-  shortId: string;
-};
-
-// Response from EventIdLookupEndpoint
-// /organizations/${orgSlug}/eventids/${eventId}/
-export type EventIdResponse = {
-  organizationSlug: string;
-  projectSlug: string;
-  groupId: string;
-  eventId: string;
-  event: Event;
-};
